Move new code to new files
